% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{and}
\alias{and}
\alias{and.operator}
\title{Applies the intersect operator on the cell types and on the cells
that participate on common cell.types}
\usage{
and.operator(hash.index1, hash.index2)
}
\arguments{
\item{hash.index1}{result index to be intersected with hash.index2}

\item{hash.index2}{result index to be intersected with hash.index1}
}
\value{
the intersected set
}
\description{
Applies the intersect operator on the cell types and on the cells
that participate on common cell.types
}
