% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{buildAnnDataCellTypeIndex}
\alias{buildAnnDataCellTypeIndex}
\alias{buildCellTypeIndex.anndata}
\title{Builds an \code{SCFind} object from a \code{AnnData} object}
\usage{
buildCellTypeIndex.anndata(path, dataset.name, cell.type.label = "cell_type1", qb = 2)
}
\arguments{
\item{path}{object of AnnData class}

\item{dataset.name}{name of the dataset that will be prepended in each cell_type}

\item{cell.type.label}{the cell.type metadata of the colData SingleCellExperiment that will be used for the index}

\item{qb}{number of bits per cell that are going to be used for quantile compression of the expression data}
}
\value{
an SCFind object
}
\description{
This function will index a \code{AnnData} as an SCFind index.
}
