% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{buildCellTypeIndex}
\alias{buildCellTypeIndex}
\alias{buildCellTypeIndex}
\alias{buildCellTypeIndex.SCESet}
\alias{buildCellTypeIndex,SingleCellExperiment,ANY,character,character-method}
\alias{buildCellTypeIndex}
\title{Build a cell type Index}
\usage{
buildCellTypeIndex(sce, dataset.name = "", assay.name = "logcounts",
  cell.type.label = "cell_type1")

buildCellTypeIndex.SCESet(sce, dataset.name, assay.name, cell.type.label)


  \S4method{buildCellTypeIndex}{SingleCellExperiment,ANY,character,character}(sce,
  dataset.name = "", assay.name = "logcounts",
  cell.type.label = "cell_type1")
}
\arguments{
\item{sce}{object of SingleCellExperiment class}

\item{dataset.name}{name of the dataset that will be prepended in each cell_type}
}
\value{
a `data.frame` containing calculated gene index
}
\description{
Calculates a fraction of expressed cells per gene per cell type
}
\examples{
TODO

}
