% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{cell.type.marker}
\alias{cell.type.marker}
\title{Find marker genes for a specific cell type
TODO(the genes can be ordered by precision recall or the f1 metric)}
\usage{
cell.type.marker(object, cell.types, background.cell.types, top.k = 5)
}
\arguments{
\item{object}{SCFind object}

\item{cell.types}{the cell types that we want to extract the marker genes}

\item{background.cell.types}{the universe of cell.types to consider}

\item{top.k}{how many genes to retrieve}
}
\value{
a data.frame that each row represent a gene score for a specific cell type
}
\description{
Find marker genes for a specific cell type
TODO(the genes can be ordered by precision recall or the f1 metric)
}
