% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{cell.type.union.operator}
\alias{cell.type.union.operator}
\title{Gets two list of gene hits results, merges cell.types and applies an operator
on the common cell types cell list}
\usage{
cell.type.union.operator(hash.index1, hash.index2, operator.function)
}
\description{
Gets two list of gene hits results, merges cell.types and applies an operator
on the common cell types cell list
}
