% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{cell.types.phyper.test}
\alias{cell.types.phyper.test}
\title{Runs a query and performs the hypergeometric test for the retrieved cell types}
\usage{
cell.types.phyper.test(object, gene.list, datasets)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.list}{the list of genes to be queried}

\item{dataset}{the datasets that will be tested as background for the hypergeometric test}
}
\value{
a DataFrame that contains all cell types with the respective cell cardinality and the hypergeometric test
}
\description{
Runs a query and performs the hypergeometric test for the retrieved cell types
}
