% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{cellTypeMarkers}
\alias{cellTypeMarkers}
\title{return all the gene markers for a specified cell.type}
\usage{
cellTypeMarkers(object, cell.types, background.cell.types, top.k = 5,
  sort.field = "f1")
}
\description{
return all the gene markers for a specified cell.type
}
