% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{cellTypeNames}
\alias{cellTypeNames}
\alias{get.cell.types.names}
\title{Return a vector with all existing cell type names in the database}
\usage{
get.cell.types.names(object, datasets)
}
\arguments{
\item{object}{SCFind object}
}
\value{
a character list
}
\description{
Return a vector with all existing cell type names in the database
}
