% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{evaluate.cell.type.markers}
\alias{evaluate.cell.type.markers}
\title{Evaluate a user specific query by calculating the precision recall metrics}
\usage{
evaluate.cell.type.markers(object, gene.list, cell.types,
  background.cell.types, sort.field)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.list}{the list of genes to be evaluated}

\item{cell.types}{a list of cell types for the list to evaluated}

\item{background.cell.types}{the universe of cell.types to consider}

\item{sort.field}{the dataframe will be sorted according to this field}
}
\value{
a DataFrame that each row represent a gene score for a specific cell type
}
\description{
Evaluate a user specific query by calculating the precision recall metrics
}
