% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{evaluatePseudotimeContinuousDensity}
\alias{evaluatePseudotimeContinuousDensity}
\alias{pseudotime.continuous.density}
\alias{evaluatePseudotimeContinuousDensity,SCFind,character-method}
\title{Calculate the density of cells expressing the gene list over a period of pseudotime}
\usage{
pseudotime.continuous.density(object, gene.list, dataset, bw = 0.1, coldata.slot = "pseudotime")

\S4method{evaluatePseudotimeContinuousDensity}{SCFind,character}(object, gene.list, dataset, bw = 0.1, coldata.slot = "pseudotime")
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.list}{genes to be searched in the gene.index}

\item{datasets}{the datasets that will be considered}

\item{bw}{the bandwidth to be used}

\item{coldata.slot}{name of the pseudotime slot in colData of the metadata}
}
\value{
the continous density of cells over a period of pseudotime
}
\description{
Evaluate the continous density of cells over a period of pseudotime
}
