% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{evaluatePseudotimeDensity}
\alias{evaluatePseudotimeDensity}
\alias{pseudotime.density}
\alias{evaluatePseudotimeDensity,SCFind,character-method}
\title{Calculate the density of cells expressing the gene list over a period of pseudotime}
\usage{
pseudotime.density(object, gene.list, dataset, specific.period = c(0, 1), coldata.slot = "pseudotime")

\S4method{evaluatePseudotimeDensity}{SCFind,character}(object, gene.list, dataset, specific.period = c(0, 1), coldata.slot = "pseudotime")
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.list}{genes to be searched in the gene.index}

\item{datasets}{the datasets that will be considered}

\item{specific.period}{set window of pseudotime considered}

\item{coldata.slot}{name of the pseudotime slot in colData of the metadata}
}
\value{
the density of cell
}
\description{
Evaluate the density of cells over a period of pseudotime
}
