% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\docType{data}
\name{ExampleIndex}
\alias{ExampleIndex}
\title{Single cell RNA-Seq datasets extracted from a publication by the Tabula Muris Consortium.}
\format{A link to the scfind index of class \code{SCFind} with 42,192 cells.}
\source{
\url{https://doi.org/10.1038/s41586-018-0590-4}

}
\usage{
ExampleIndex
}
\description{
A preprocessed scfind index of the Tabula Muris (FACS) & Tabula Muris (10X)
}
\keyword{datasets}