% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\docType{data}
\name{ExampleIndex_TabulaMuris10X}
\alias{ExampleIndex_TabulaMuris10X}
\title{Single cell RNA-Seq data extracted from a publication by the Tabula Muris Consortium.}
\format{A link to the scfind index of class \code{SCFind} with 54,837 cells.}
\source{
\url{https://doi.org/10.1038/s41586-018-0590-4}

}
\usage{
ExampleIndex_TabulaMuris10X
}
\description{
A preprocessed scfind index of the Tabula Muris (10X)
}
\keyword{datasets}