% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{find.marker.genes}
\alias{find.marker.genes}
\title{Query Optimization Function for SCFind objects.}
\usage{
find.marker.genes(object, gene.list, datasets)
}
\arguments{
\item{object}{SCFind object}

\item{gene.list}{A list of Genes existing in the database}

\item{datasets}{the datasets of the objects to be considered}
}
\value{
hierarchical list of queries and their respective scores
}
\description{
This function can be used with quite long gene lists
that otherwise would have no cell hits in the database
}
