% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{findCell}
\alias{findCell}
\alias{findCell}
\alias{findCell.SCESet}
\alias{findCell,list-method}
\alias{findCell}
\title{Find cells associated with a given gene list}
\usage{
findCell(input = NULL, genelist = NULL)

findCell.SCESet(input, genelist)

\S4method{findCell}{list}(input = NULL, genelist = NULL)
}
\arguments{
\item{input}{object of SingleCellExperiment class}

\item{genelist}{column name in the colData slot of the object SingleCellExperiment 
containing the cell classification information}
}
\value{
a `list` containing calculated gene index
}
\description{
Calculates p-values of a log-likelihood of a list of genes to be associated
with each cell type. Log-likelihood is based on gene expression values.
}
\examples{
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays = list(normcounts = as.matrix(yan)), colData = ann)
# this is needed to calculate dropout rate for feature selection
# important: normcounts have the same zeros as raw counts (fpkm)
counts(sce) <- normcounts(sce)
logcounts(sce) <- log2(normcounts(sce) + 1)
# use gene names as feature symbols
rowData(sce)$feature_symbol <- rownames(sce)
isSpike(sce, 'ERCC') <- grepl('^ERCC-', rownames(sce))
# remove features with duplicated names
sce <- sce[!duplicated(rownames(sce)), ]
index <- buildCellIndex(sce)
res <- findCell(index, genelist = c('SOX6', 'SNAI3'))

}
