% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\name{findCellTypes}
\alias{findCellTypes}
\alias{findCellTypes.geneList}
\title{Find cell types associated with a given gene list. All cells
returned express all of the genes in the given gene list}
\usage{
findCellTypes(object, gene.list, datasets)

findCellTypes.geneList(object, gene.list, datasets)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.list}{genes to be searched in the gene.index}
}
\value{
a named numeric vector containing p-values
}
\description{
Find cell types associated with a given gene list. All cells
returned express all of the genes in the given gene list
}
\examples{
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays = list(normcounts = as.matrix(yan)), colData = ann)
# this is needed to calculate dropout rate for feature selection
# important: normcounts have the same zeros as raw counts (fpkm)
counts(sce) <- normcounts(sce)
logcounts(sce) <- log2(normcounts(sce) + 1)
# use gene names as feature symbols
rowData(sce)$feature_symbol <- rownames(sce)
isSpike(sce, 'ERCC') <- grepl('^ERCC-', rownames(sce))
# remove features with duplicated names
sce <- sce[!duplicated(rownames(sce)), ]
index <- buildCellIndex(sce)
res <- findCell(index, genelist = c('SOX6', 'SNAI3'))

}
