% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{findCellTypeSpecificities}
\alias{findCellTypeSpecificities}
\alias{cell.type.specificity}
\alias{findCellTypeSpecificities,SCFind,character-method}
\title{Find out how many cell-types each gene is found}
\usage{
findCellTypeSpecificities(object, gene.list, datasets, min.cells=10, min.fraction=.25)

\S4method{findCellTypeSpecificities}{SCFind,character}(object, gene.list, datasets, min.cells, min.fraction)
}
\arguments{
\item{object}{SCFind object}

\item{gene.list}{A list of Genes existing in the database}

\item{datasets}{the datasets that will be considered}

\item{min.cells}{minimum threshold}

\item{min.fraction}{percentage of total cell in a cell type as threshold}
}
\value{
list of genes and their number of cell types
}
\description{
This function can be used with quite long gene lists
that otherwise would have no cell hits in the database
}
