% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{findGeneSignatures}
\alias{findGeneSignatures}
\alias{gene.signatures}
\alias{findGeneSignatures,SCFind,character-method}
\title{Find the signature genes for a cell-type}
\usage{
findGeneSignatures(object, cell.types, max.genes=1000, min.cells=10, max.pval=0) 

\S4method{findGeneSignatures}{SCFind,character}(object, cell.types, max.genes, min.cells, max.pval) 
}
\arguments{
\item{object}{SCFind object}

\item{cell.types}{the cell types that we want to extract the house keeping genes}

\item{max.genes}{maximum threshold of number of genes to be considered}

\item{min.cells}{minimum threshold of cell hit}

\item{max.pval}{maximum threshold of p-value}
}
\value{
list of genes that are signatures in the query of each cell type
}
\description{
by ranking genes by recall and then adding genes to the query until we exceed a target p-value threshold or until a minimum number of cells is returned from the query
}
