% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{findHouseKeepingGenes}
\alias{findHouseKeepingGenes}
\alias{house.keeping.genes}
\alias{findHouseKeepingGenes,SCFind,character-method}
\title{Find the set of genes that are ubiquitously expressed in a query of cell types}
\usage{
findHouseKeepingGenes(object, cell.types, min.recall=.5, max.genes=1000)

\S4method{findHouseKeepingGenes}{SCFind,character}(object, cell.types, min.recall, max.genes)
}
\arguments{
\item{object}{SCFind object}

\item{cell.types}{the cell types that we want to extract the house keeping genes}

\item{min.recall}{minimum threshold of recall value}

\item{max.genes}{maximum threshold of number of genes to be considered}
}
\value{
list of genes that are ubiquitously expressed in the query of cell types
}
\description{
This function can be used with quite long cell type list
}
