% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{findSimilarGenes}
\alias{findSimilarGenes}
\alias{similar.genes}
\alias{findSimilarGenes,SCFind,character-method}
\title{Look at all other genes and rank them based on the similarity of their expression pattern to the pattern defined by the gene query}
\usage{
findSimilarGenes(object, gene.list, datasets, top.k = 5)

\S4method{findSimilarGenes}{SCFind,character}(object, gene.list, datasets, top.k)
}
\arguments{
\item{object}{SCFind object}

\item{gene.list}{A list of Genes existing in the database}

\item{datasets}{the datasets that will be considered}

\item{top.k}{how many genes to retrieve}
}
\value{
list of genes and their similarities presented in Jaccard indices
}
\description{
Look at all other genes and rank them based on the similarity of their expression pattern to the pattern defined by the gene query
}
