% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{findTissueSpecificities}
\alias{findTissueSpecificities}
\alias{tissue.specificity}
\alias{findTissueSpecificities,SCFind,character-method}
\title{Find out how many cell-types each gene is found}
\usage{
findTissueSpecificities(object, gene.list, min.cells=10)

\S4method{findTissueSpecificities}{SCFind,character}(object, gene.list, min.cells)
}
\arguments{
\item{object}{SCFind object}

\item{gene.list}{A list of Genes existing in the database}

\item{min.cells}{minimum threshold}

}
\value{
list of genes and their number of tissues
}
\description{
This function can be used with quite long gene lists
that otherwise would have no cell hits in the database
}
