% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{hyperQueryCellTypes}
\alias{hyperQueryCellTypes}
\alias{cell.types.phyper.test}
\alias{hyperQueryCellTypes,SCFind,character-method}
\title{Runs a query and performs the hypergeometric test for the retrieved cell types}
\usage{
cell.types.phyper.test(object, gene.list, datasets)

\S4method{hyperQueryCellTypes}{SCFind,character}(object, gene.list,
  datasets)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.list}{the list of genes to be queried}

\item{datasets}{the datasets vector that will be tested as background for the hypergeometric test}
}
\value{
a DataFrame that contains all cell types with the respective cell cardinality and the hypergeometric test
}
\description{
Runs a query and performs the hypergeometric test for the retrieved cell types
}
