% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{importSCEMetadata}
\alias{importSCEMetadata}
\alias{sce.metadata.to.scfind.index}
\title{Import metadata to \code{SCFind} object from a \code{SingleCellExperiment} object}
\usage{
sce.metadata.to.scfind.index(sce, object, metadata , used.cell.type.label, dataset.name, index.name)

importSCEMetadata(sce, object, metadata , used.cell.type.label, dataset.name, index.name)

\S4method{importSCEMetadata}{SingleCellExperiment}(sce, object, metadata = c("colData", "reducedDims"), used.cell.type.label = "cell_type1", dataset.name, index.name)
}
\arguments{

\item{sce}{object of SingleCellExperiment class}

\item{object}{an SCFind object}

\item{metadata}{attributes of the SingleCellExperiment class that will be imported}

\item{used.cell.type.label}{the cell.type metadata of the colData SingleCellExperiment that is used for the index}

\item{dataset.name}{name of the dataset that the metadata is imported from the SingleCellExperiment class}

\item{index.name}{name of the index}
}
\value{
an SCFind object
}
\description{
This function will return an SCFind index with metadata imported from SCE class.
}
