% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\name{loadObject}
\alias{loadObject}
\alias{load.serialized.object}
\alias{loadObject,ANY-method}
\title{Generic to be used instead of readRDS}
\usage{
loadObject(filename)

load.serialized.object(filename)

\S4method{loadObject}{ANY}(filename)
}
\arguments{
\item{filename}{the filepath of a specialized serialized scfind object}
}
\value{
an \code{SCFind} object
}
\description{
After loading the database it clears the loaded bytestream from the memory.
}
