% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{markerGenes}
\alias{markerGenes}
\alias{markerGenes,SCFind,character-method}
\title{Performs query optimization and return the best candidate gene sets}
\usage{
markerGenes(object, gene.list, datasets)

\S4method{markerGenes}{SCFind,character}(object, gene.list, datasets)
}
\description{
Performs query optimization and return the best candidate gene sets
}
