% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{mergeDataset}
\alias{mergeDataset}
\alias{merge.dataset.from.object}
\alias{mergeDataset,SCFind,SCFind-method}
\alias{mergeObjects}
\title{Merges an external index into the existing object}
\usage{
\method{merge}{dataset.from.object}(object, new.object)

\S4method{mergeDataset}{SCFind,SCFind}(object, new.object)
}
\arguments{
\item{object}{the root scfind object}

\item{new.object}{external scfind object to be merged}
}
\value{
the new extended object
}
\description{
This function is useful to merge \code{SCFind} indices.
After this operation object that was merged can be discarded.
}
\details{
The only semantic limitation for merging two databases is to
have different dataset names in the two different indices.
If that is not case user may run into problems masking datasets
from the different datasets while there is a possibility of having
different cell types under the same name. This will most likely cause
undefined behavior during queries.
}
