% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{mergeIndices}
\alias{mergeIndices}
\title{merge two elias fano indices}
\usage{
mergeIndices(efdb.root, efdb)
}
\arguments{
\item{efdb.root}{the root index}

\item{efdb}{the index to be merged}
}
\description{
merge two elias fano indices
}
