% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{mergeSCE}
\alias{mergeSCE}
\alias{merge.dataset.from.sce}
\alias{mergeSCE,SCFind,SingleCellExperiment,character-method}
\title{Merges a SingleCellExperiment object into the SCFind index}
\usage{
\method{merge}{dataset.from.sce}(object, sce, dataset.name)

\S4method{mergeSCE}{SCFind,SingleCellExperiment,character}(object, sce, dataset.name)
}
\arguments{
\item{object}{the root scfind object}

\item{sce}{the \code{SingleCellExperiment} object to be merged}

\item{dataset.name}{a dataset name for the assay}
}
\value{
the new object with the sce object merged
}
\description{
It creates an \code{SCFind} for the individual assay and then invokes
the \code{mergeDataset} method obeying the same semantic rules.
}
