% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{not}
\alias{not}
\alias{not.operator}
\title{Remove cells that have some specific gene hits that we do not want in our list}
\usage{
not.operator(hash.index, hash.diff)
}
\arguments{
\item{hash.index}{the result index that cell types will be filtered}

\item{hash.diff}{the cell types to be removed in case there is an overlap}
}
\value{
an new filtered index
}
\description{
Remove cells that have some specific gene hits that we do not want in our list
}
