% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{or}
\alias{or}
\alias{or.operator}
\title{Applies the union operator on two result data structures}
\usage{
or.operator(hash.index1, hash.index2)
}
\arguments{
\item{hash.index1}{result index to be unionized with hash.index2}

\item{hash.index2}{result index to be unionized with hash.index1}
}
\value{
the unionized set
}
\description{
Applies the union operator on two result data structures
}
