% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\name{saveObject}
\alias{saveObject}
\alias{save.serialized.object}
\alias{saveObject,ANY-method}
\title{Generic to be used instead of saveRDS}
\usage{
saveObject(object, file)

save.serialized.object(object, file)

\S4method{saveObject}{ANY}(object, file)
}
\arguments{
\item{object}{an SCFind object}

\item{file}{the target filename that the object is going to be stored}
}
\value{
the \code{SCFind} object
}
\description{
This function can be used to enable the user save the loaded file in a database
to avoid re-indexing and re-merging individual assays.
}
\details{
After serializing and saving it clears the redundant bytestream from memory
because the memory is already loaded in memory
}
