% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyMethods.R
\docType{methods}
\name{scfind.get.genes.in.db}
\alias{scfind.get.genes.in.db}
\alias{scfindGenes,SCFind-method}
\title{Get all genes in the database}
\usage{
scfind.get.genes.in.db(object)

\S4method{scfindGenes}{SCFind}(object)
}
\description{
Get all genes in the database
}
