% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfind-package.R
\docType{package}
\name{scfind}
\alias{scfind}
\alias{scfind-package}
\title{Minimal Rcpp Module Example}
\description{
scfind
}
\details{
A minimal example of using Rcpp modules in an Rcpp package
}
\author{
Nikolaos Patikas
}
