% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{scfindGenes,SCFind-method}
\alias{scfindGenes,SCFind-method}
\alias{scfindGenes}
\title{Get all genes in the database}
\usage{
\S4method{scfindGenes}{SCFind}(object)
}
\description{
Get all genes in the database
}
