% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{scfindQ2loadDictionaries}
\alias{scfindQ2loadDictionaries}
\alias{read_all_dictionaries}
\alias{scfindQ2loadDictionaries,ANY,character-method}
\title{Load up dictionaries for free text search function}
\usage{
read_all_dictionaries(w2v, dictionary, priority)

\S4method{scfindQ2loadDictionaries}{ANY,character}(w2v, dictionary, priority)
}
\arguments{
\item{w2v}{the word2vec model}

\item{dictionary}{valid dictionary generated from hemberg-lab}

\item{priority}{setting priority of dictionary}
}
\value{
the list of dictionaries
}
\description{
Load up dictionaries for free text search function
}
