% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyMethods.R
\docType{methods}
\name{scfindShinyServer}
\alias{scfindShinyServer}
\alias{server.scfind}
\alias{scfindShinyServer,SCFind-method}
\title{Server handler for scfind}
\usage{
server.scfind(object)

\S4method{scfindShinyServer}{SCFind}(object)
}
\arguments{
\item{object}{An SCFind object that the shiny app will be deployed upon}
}
\description{
Server handler for scfind
}
