% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyMethods.R
\docType{methods}
\name{scfindShinyW2VServer}
\alias{scfindShinyW2VServer}
\alias{server.scfind.w2v}
\alias{scfindShinyW2VServer,SCFind-method}
\title{Server handler for free text search scfind}
\usage{
server.scfind(object, dictionary)

\S4method{scfindShinyW2VServer}{SCFind}(object, dictionary)
}
\arguments{
\item{object}{An SCFind object that the shiny app will be deployed upon}
}
\description{
Server handler for free text search scfind
}
