% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyMethods.R
\name{server.scfind}
\alias{server.scfind}
\title{Server handler for scfind}
\usage{
server.scfind(object)
}
\arguments{
\item{object}{An SCFind object that the shiny app will be deployed upon}
}
\description{
Server handler for scfind
}
