% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm10x.R
\docType{data}
\name{tm10x}
\alias{tm10x}
\title{`tm10x` SCE objects created from The Tabula Muris Consortium.}
\format{
list of url to download SCE objects.
}
\source{
\href{https://figshare.com/articles/Single-cell_RNA-seq_data_from_Smart-seq2_sequencing_of_FACS_sorted_cells/5715040}{figshare}
}
\usage{
data(tm10x)
}
\description{
SCE objects created from The Tabula Muris Consortium. https://doi.org/10.1038/s41586-018-0590-4
that profiled every cell-type in the mouse using high-coverage FACS-sorted cells + Smartseq2 in the original publication
}
\references{
The Tabula Muris Consortium. (2018) 
Single-cell transcriptomics of 20 mouse
organs creates a Tabula Muris.
Nature Oct;562(7727):367-372.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/30283141}{PubMed})
}
\keyword{datasets}
