% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\docType{data}
\name{tmfacs}
\alias{tmfacs}
\title{Single cell RNA-Seq data extracted from a publication by the Tabula Muris Consortium.}
\format{A link to the scfind index of class \code{SingleCellExperiment} with 42,192 cells.}
\source{
\url{https://doi.org/10.1038/s41586-018-0590-4}

}
\usage{
tmfacs
}
\description{
A preprocessed scfind SCE set object of the Tabula Muris (FACS)
}
\keyword{datasets}