% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyMethods.R
\name{ui.scfind}
\alias{ui.scfind}
\title{UI handler for the shiny front end of scfind}
\usage{
ui.scfind(object)
}
\arguments{
\item{object}{SCFind object}
}
\description{
UI handler for the shiny front end of scfind
}
