% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\name{cellTypeMarkers}
\alias{cellTypeMarkers}
\alias{cell.type.marker}
\alias{cellTypeMarkers,SCFind,character-method}
\title{return all the gene markers for a specified cell.type}
\usage{
cellTypeMarkers(
  object,
  cell.types,
  background.cell.types,
  top.k = 5,
  sort.field = "f1"
)

cell.type.marker(object, cell.types, background.cell.types, top.k, sort.field)

\S4method{cellTypeMarkers}{SCFind,character}(
  object,
  cell.types,
  background.cell.types,
  top.k = 5,
  sort.field = "f1"
)
}
\arguments{
\item{object}{SCFind object}

\item{cell.types}{the cell types that we want to extract the marker genes}

\item{background.cell.types}{the universe of cell.types to consider}

\item{top.k}{how many genes to retrieve}

\item{sort.field}{the dataframe will be sorted according to this field}
}
\value{
a data.frame that each row represent a gene score for a specific cell type
}
\description{
return all the gene markers for a specified cell.type

Find marker genes for a specific cell type
}
