% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{findCellTypes}
\alias{findCellTypes}
\alias{findCellTypes.geneList}
\alias{findCellTypes,SCFind,character-method}
\title{queries cells that contain all the genes from the list}
\usage{
findCellTypes(object, gene.list, datasets)

findCellTypes.geneList(object, gene.list, datasets)

\S4method{findCellTypes}{SCFind,character}(object, gene.list, datasets)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.list}{genes to be searched in the gene.index}

\item{datasets}{the datasets that will be considered}
}
\value{
a named numeric vector containing p-values
}
\description{
queries cells that contain all the genes from the list

Find cell types associated with a given gene list. All cells
returned express all of the genes in the given gene list

queries cells that contain all the genes from the list
}
