% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\name{findCellTypeSpecificities}
\alias{findCellTypeSpecificities}
\alias{cell.type.specificity}
\alias{findCellTypeSpecificities,SCFind-method}
\title{Find out how many cell-types each gene is found}
\usage{
findCellTypeSpecificities(
  object,
  gene.list,
  datasets,
  min.cells = 10,
  min.fraction = 0.25
)

cell.type.specificity(
  object,
  gene.list,
  datasets,
  min.cells = 10,
  min.fraction = 0.25
)

\S4method{findCellTypeSpecificities}{SCFind}(
  object,
  gene.list,
  datasets,
  min.cells = 10,
  min.fraction = 0.25
)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.list}{genes to be searched in the gene.index}

\item{datasets}{the datasets that will be considered}

\item{min.cells}{threshold of cell hit of a cell type}

\item{min.fraction}{portion of total cell as threshold}
}
\value{
the list of number of cell type for each gene
}
\description{
Find out how many cell-types each gene is found

Find out how many cell-types each gene is found
}
