% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\name{findTissueSpecificities}
\alias{findTissueSpecificities}
\alias{tissue.specificity}
\alias{findTissueSpecificities,SCFind-method}
\title{Find out how many tissues each gene is found}
\usage{
findTissueSpecificities(object, gene.list, min.cells = 10)

tissue.specificity(object, gene.list, min.cells = 10)

\S4method{findTissueSpecificities}{SCFind}(object, gene.list, min.cells = 10)
}
\arguments{
\item{object}{the \code{SCFind} object}

\item{gene.list}{genes to be searched in the gene.index}

\item{min.cells}{threshold of cell hit of a tissue}
}
\value{
the list of number of tissue for each gene
}
\description{
Find out how many tissues each gene is found

Find out how many tissues each gene is found
}
