% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{markerGenes}
\alias{markerGenes}
\alias{find.marker.genes}
\alias{markerGenes,SCFind,character,character-method}
\title{Performs query optimization and return the best candidate gene sets}
\usage{
markerGenes(object, gene.list, datasets)

find.marker.genes(object, gene.list, datasets)

\S4method{markerGenes}{SCFind,character,character}(object, gene.list,
  datasets)
}
\arguments{
\item{object}{SCFind object}

\item{gene.list}{A list of Genes existing in the database}

\item{datasets}{the datasets of the objects to be considered}
}
\value{
hierarchical list of queries and their respective scores
}
\description{
This function can be used with quite long gene lists
that otherwise would have no cell hits in the database
}
