% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyMethods.R
\docType{methods}
\name{scfind.interactive}
\alias{scfind.interactive}
\alias{scfindShiny,SCFind-method}
\title{Opens \code{scfind} index in an interactive session in a web browser.}
\usage{
scfind.interactive(object)

\S4method{scfindShiny}{SCFind}(object)
}
\arguments{
\item{object}{an object of \code{SCFind} class}
}
\value{
Opens a browser window with an interactive \code{shiny} app and visualize queries on the dataset.
}
\description{
Runs interactive \code{shiny} session of \code{scfind} based on the indexed project.
}
