% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{scfindGenes}
\alias{scfindGenes}
\alias{scfind.get.genes.in.db}
\alias{scfindGenes,SCFind-method}
\title{Get all genes in the database}
\usage{
scfind.get.genes.in.db(object)

\S4method{scfindGenes}{SCFind}(object)
}
\arguments{
\item{object}{the \code{scfind} object}
}
\value{
the list of genes present in the database
}
\description{
Get all genes in the database
}
