% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{scfindQ2loadDictionaries}
\alias{scfindQ2loadDictionaries}
\alias{read_all_dictionaries}
\title{load up dictionaries for free text search}
\usage{
scfindQ2loadDictionaries(w2v, dictionary)

}
\arguments{
\item{w2v}{the filepath of a word2vec model bin file}
\item{dictionary}{the filepath of a valid dictionary from the hemberg-lab}
}
\description{
Dictionaries for free text search
}
