% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyMethods.R
\docType{methods}
\name{scfindShinyServer,SCFind-method}
\alias{scfindShinyServer,SCFind-method}
\alias{scfindShinyServer}
\title{The scfind server method}
\usage{
\S4method{scfindShinyServer}{SCFind}(object)
}
\description{
The scfind server method
}
