% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyMethods.R
\name{server.scfind}
\alias{server.scfind}
\title{Server handler for scfind}
\usage{
server.scfind(object)
}
\arguments{
\item{input}{handler item for the ShinyApp}

\item{output}{handel item for the ShinyApp}

\item{session}{handler item for the stateful ShinyApp}
}
\description{
Server handler for scfind
}
